Specify ruby version used, both for ruby version managers like rvm, chruby, rbenv as well as heroku.
Why not only in Gemfile? Because that does not work with patchlevel specifications,
see http://bundler.io/v1.3/gemfile_ruby.html. Having a .ruby-version file with an
explicit patch level will force developers to use the latest patchlevel defined in
the repository, instead of running on any sufficiently qualified version.

https://gist.github.com/fnichol/1912050
https://devcenter.heroku.com/articles/ruby-versions