Finally, an actual rack application, bootstrapped with the cookie session store middleware to demonstrate the usage of both rack middleware and configuring it using our ENV variables.
To keep our config.ru as simple and logic-free as possible and to allow testing
our components as well as the routing itself later on, we simply define a rack
application at the root level constant ::Routes in config/routes.rb.

The routes.rb file uses Rack::Builder by default to demonstrate composition
of rack applications from various endpoints and mixing in of middleware in this stack.